requirements.txt
python-decouple==3.8
schedule==1.2.2
